[FRONTEND] Moved the logic out of the tsx component
